+++ push_swap/libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: taeskim <taeskim@student.42.fr>            +#+  +:+       +#+        */
+/*   By: pac-man <pac-man@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/14 13:33:23 by taeskim           #+#    #+#             */
-/*   Updated: 2021/06/15 11:28:20 by taeskim          ###   ########.fr       */
+/*   Updated: 2021/07/04 20:22:58 by pac-man          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 # define LIBFT_H
 # include <stdio.h>
 
-int		ft_isdigit(char c);
-int		ft_atoi(const char *src);
-int		ft_strlen(char *s);
+int				ft_isdigit(char c);
+long long		ft_atoi(const char *src);
+int				ft_strlen(char *s);
 
 
 #endif
